starknet: move 'calldata' to 'core' directory and make it more maintainable
