+++ packages/starknet/lib/src/contract/index.dart
@@ -1,3 +1,3 @@
-export 'model/index.dart';
 export 'contract.dart';
 export 'erc20.dart';
+export 'model/index.dart';
