+++ packages/starknet/lib/src/core/index.dart
@@ -1,3 +1,4 @@
+export 'calldata.dart';
 export 'convert.dart';
 export 'signer/index.dart';
 export 'snip/index.dart';
