+++ packages/starknet/lib/starknet.dart
@@ -6,7 +6,6 @@ library starknet;
 export 'src/account.dart';
 export 'src/account/index.dart';
 export 'src/argent/index.dart';
-export 'src/calldata.dart';
 export 'src/contract/index.dart';
 export 'src/core/index.dart';
 export 'src/crypto/index.dart';
